% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{compute_umap}
\alias{compute_umap}
\title{Compute UMAPs.}
\usage{
compute_umap(df, num_comp = 2, subset = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{num_comp}{A positive integer. The number of UMAP components to compute.}

\item{subset}{A character vector. The subset of columns to use for UMAP
calculation. Default is all columns. Columns in \code{df} outside of this subset
will be kept in the result.}
}
\value{
A data frame with the UMAP components.
}
\description{
Computation is be done using the \code{embed} package.
}
\examples{
if (rlang::is_installed("mirmisc")) {
  ga_data <- get_ga_data(
    log2 = TRUE,
    gene_predicate = ~ median(.) > 0
  )
  genes <- dplyr::intersect(names(ga_data), mirmisc::get_gene_names())
  umap <- compute_umap(ga_data, subset = genes)
  ggplot2::ggplot(
    umap,
    ggplot2::aes(umap_1, umap_2, color = meta_collectionga)
  ) +
    ggplot2::geom_point() +
    ggplot2::scale_color_viridis_c()
}
}
