% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conv_traintest_lst_to_rsplit}
\alias{conv_traintest_lst_to_rsplit}
\title{Convert a list of \code{training_data} and \code{testing_data} to an \code{rsplit}.}
\usage{
conv_traintest_lst_to_rsplit(traintest_lst)
}
\arguments{
\item{traintest_lst}{A list with two elements \code{training_data} and
\code{testing_data} which are data frames with the same column names.}
}
\value{
An \code{rsplit} object. Something like the return of
\code{\link[rsample:initial_split]{rsample::initial_split()}}.
}
\description{
A lot of functions in this package return lists with two elements  that are
data frames called \code{training_data} and \code{testing_data}. You might prefer these
to be in the form of an \code{rsplit} object, such as the return of
\code{\link[rsample:initial_split]{rsample::initial_split()}} so that you can access the training and testing
parts with \code{\link[rsample:initial_split]{rsample::training()}} and \code{\link[rsample:initial_split]{rsample::testing()}}.
}
