% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential-expression.R
\name{cor_de}
\alias{cor_de}
\title{Conduct a differential expression analysis using correlation tests.}
\usage{
cor_de(
  data,
  condition,
  genes = NULL,
  method = c("spearman", "kendall", "pearson"),
  padj_method = stats::p.adjust.methods[1]
)
}
\arguments{
\item{data}{A data frame where rows are samples and columns are genes and
metadata.}

\item{condition}{A string. The name of the column in \code{data} which contains
the condition on which the differential expression is premised. This column
should be logical, a factor with two levels or a numeric vector. The first
level (or \code{FALSE}) is the baseline and the second level (or \code{TRUE}) is the
disease/treatment. If the condition is numeric, then columns \code{log2fc},
\code{base_med}, \code{case_med}, \code{base_mean} and \code{case_mean} will be absent from the
result.}

\item{genes}{A character vector. The column names that contain the genes for
the differential expression. If \code{NULL}, any genes in data found in
\code{\link[mirmisc:get_gene_names]{mirmisc::get_gene_names()}} are used.}

\item{method}{A string. The correlation test method. Must be \code{"spearman"},
\code{"kendall"} or \code{"pearson"}.}

\item{padj_method}{A string. The method of adjusting the p-values for
multiple hypothesis testing. Must be one of \code{\link[stats:p.adjust]{stats::p.adjust.methods()}}.}
}
\value{
A tibble with 8 columns:
\itemize{
\item \code{gene}: The gene name.
\item \code{log2fc}: The log2 fold-change between case and control.
\item \code{pvalue}: The p-value for that gene being differentially expressed.
\item \code{padj}: Adjusted p-values.
\item \code{base_med}: The median raw value for the baseline samples.
\item \code{case_med}: The median raw value for the case samples.
\item \code{base_mean}: The mean raw value for the baseline samples.
\item \code{case_mean}: The mean raw value for the case samples. For \code{method = "pearson"} the \code{log2fc} is of the
means of the two groups. Otherwise, it is of the medians.
}
}
\description{
This function is designed to ingest data in the form output by functions like
\code{\link[=get_bw_data]{get_bw_data()}}.
}
\details{
This function has no way of dealing with batch effects. To allow for these,
you should correct your data (with e.g. \code{\link[=linear_correct]{linear_correct()}})
beforehand.
}
\examples{
if (requireNamespace("mirmisc", quietly = TRUE)) {
  bwms_data <- get_combined_cohort_data(
    c("bw", "ms"),
    log2 = TRUE, tot_counts = TRUE,
    gene_predicate = ~ stats::median(.) > 0,
  ) \%>\%
    dplyr::filter(!is.na(meta_pre_eclampsia)) \%>\%
    dplyr::mutate(tot_counts = log2(tot_counts + 1))
  genes <- dplyr::intersect(mirmisc::get_gene_names(), names(bwms_data))
  bwms_corr <- linear_correct(
    bwms_data,
    correct_cols = genes,
    correct_for_cols = c("log2_tot_counts", "cohort",
                         "meta_q_pcr_actb_ct", "meta_q_pcr_ercc_ct"),
    keep_effect_cols = "meta_pre_eclampsia"
  )[[1]]
  cor_de(bwms_corr, "meta_pre_eclampsia")
}
}
\seealso{
Other differential expression methods: 
\code{\link{deseq}()},
\code{\link{edger}()}
}
\concept{differential expression methods}
