% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential-expression.R
\name{edger}
\alias{edger}
\title{Conduct a differential expression analysis using \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.}
\usage{
edger(
  data,
  condition,
  batch = NULL,
  genes = NULL,
  quick = FALSE,
  padj_method = stats::p.adjust.methods[1]
)
}
\arguments{
\item{data}{A data frame where rows are samples and columns are genes and
metadata.}

\item{condition}{A string. The name of the column in \code{data} which contains
the condition on which the differential expression is premised. This column
should be logical or a factor with two levels. The first level (or \code{FALSE})
is the baseline and the second level (or \code{TRUE}) is the disease/treatment.}

\item{batch}{A string. The name of the column in \code{data} specifying the
batches that you'd like the model to account for. A common choice is a
cohort column.}

\item{genes}{A character vector. The column names that contain the genes for
the differential expression. If \code{NULL}, any genes in data found in
\code{\link[mirmisc:get_gene_names]{mirmisc::get_gene_names()}} are used.}

\item{quick}{A flag. If \code{TRUE}, the \code{base_med}, \code{case_med}, \code{base_mean} and
\code{case_mean} columns are omitted from the return and the lack of need to
calculate these medians offers a small speedup.}

\item{padj_method}{A string. The method of adjusting the p-values for
multiple hypothesis testing. Must be one of \code{\link[stats:p.adjust]{stats::p.adjust.methods()}}.}
}
\value{
A tibble with 8 columns:
\itemize{
\item \code{gene}: The gene name.
\item \code{log2fc}: The log2 fold-change between case and control.
\item \code{pvalue}: The p-value for that gene being differentially expressed.
\item \code{padj}: Adjusted p-values.
\item \code{base_med}: The median raw value for the baseline samples.
\item \code{case_med}: The median raw value for the case samples.
\item \code{base_mean}: The mean raw value for the baseline samples.
\item \code{case_mean}: The mean raw value for the case samples.
}
}
\description{
This function is designed to ingest data in the form output by functions like
\code{\link[=get_bw_data]{get_bw_data()}}.
}
\examples{
bw_data <- get_bw_data()
edger(bw_data, "meta_pre_eclampsia")
}
\seealso{
Other differential expression methods: 
\code{\link{cor_de}()},
\code{\link{deseq}()}
}
\concept{differential expression methods}
