% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-lm.R
\name{multi_lm}
\alias{multi_lm}
\title{Fit multiple linear models.}
\usage{
multi_lm(df, xs, ys, robust = TRUE)
}
\arguments{
\item{df}{A data frame containing explanatory variables \code{xs} and dependent
variables \code{ys}.}

\item{xs}{A character vector. The explanatory variables. These must be the
names of columns of \code{df} that are either numeric, factor or logical.}

\item{ys}{A character vector. The dependent variables. These must be the
names of numeric columns of \code{df} that are either numeric or factors.}

\item{robust}{A flag. Use robust linear model \code{\link[MASS:rlm]{MASS::rlm()}}? Can only be used
with \code{type = 1}.}
}
\value{
A list of fitted linear models (the results of calls to \code{\link[stats:lm]{stats::lm()}} or \code{\link[MASS:rlm]{MASS::rlm()}}).
}
\description{
For a constant set of explanatory variables (\code{xs}) and several dependent
variables (\code{ys}), fit a linear model \code{y ~ xs}  for each \code{y} in \code{ys}.
}
\examples{
if (rlang::is_installed("mirmisc")) {
  gars_data <- get_combined_cohort_data(c("ga", "rs"),
    gene_predicate = ~ median(.) > 0, log2 = TRUE
  ) \%>\%
    dplyr::mutate(
      cohort = factor(
        dplyr::if_else(startsWith(mirvie_id, "RS"), "RS", "GA")
      )
    ) \%>\%
    dplyr::filter(!is.na(meta_major_race), !is.na(meta_collectionga))
  xs <- c("cohort", "meta_major_race", "meta_collectionga")
  ys <- mirmisc::get_df_gene_names(gars_data)
  res <- multi_lm(gars_data, xs, ys)
}
}
