% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curve.R
\name{new_mirvie_learning_curve}
\alias{new_mirvie_learning_curve}
\title{Construct a \code{mirvie_learning_curve} object.}
\usage{
new_mirvie_learning_curve(tib)
}
\arguments{
\item{tib}{A tibble with columns \code{rep} (int), \code{testing_frac} (dbl),
\code{training_frac} (dbl), \code{testing_indices} (list), \code{training_indices} (list),
\code{cv} (dbl) and \code{test} (dbl).}
}
\value{
A \code{mirvie_learning_curve} object.
}
\description{
A \code{mirvie_learning_curve} object is what is output by \code{\link[=learn_curve]{learn_curve()}}.
}
\details{
This just tacks \code{"mirvie_learning_curve"} onto the front of the \code{class}
attribute of an appropriate object. This should only be used inside of
\code{\link[=learn_curve]{learn_curve()}}.
}
