% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curve.R
\name{new_mirvie_learning_curve_cv}
\alias{new_mirvie_learning_curve_cv}
\title{Construct a \code{mirvie_learning_curve} object.}
\usage{
new_mirvie_learning_curve_cv(tib)
}
\arguments{
\item{tib}{A tibble with columns \code{training_samples} (integer, the number of
training samples on that iteration) and \code{score} (double, a model metric
score).}
}
\value{
A \code{mirvie_learning_curve_cv} object.
}
\description{
A \code{mirvie_learning_curve_cv} object is what is output by \code{\link[=learn_curve_cv]{learn_curve_cv()}}.
}
\details{
This just tacks \code{"mirvie_learning_curve_cv"} onto the front of the \code{class}
attribute of an appropriate object. This should only be used inside of
\code{\link[=learn_curve_cv]{learn_curve_cv()}}.
}
