% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot_pvals}
\alias{qqplot_pvals}
\title{QQ-plot a vector of p-values.}
\usage{
qqplot_pvals(
  x,
  labels = NULL,
  controls = NULL,
  cases = NULL,
  lambda_correct = FALSE,
  aes_opts = list(null_alpha = 1, fit_alpha = 1)
)
}
\arguments{
\item{x}{A numeric vector with elements in (0, 1).}

\item{labels}{A character vector of length at most \code{length(x)}. Labels for
the smallest p-values.}

\item{controls}{Number of controls in the tests.}

\item{cases}{Number of cases in the tests.}

\item{lambda_correct}{A flag. Apply a correction to make the slope of the
fitted line equal to 1.}

\item{aes_opts}{A list. Aesthetic options for the plot. The list has two
named elements.
\itemize{
\item \code{null_alpha}: Transparency of the \code{x = y} line.
\item \code{fit_alpha}: Transparency of the fit line.
}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
QQ-plot a vector of p-values.
}
\examples{
set.seed(1)
pvals <- runif(999) ^ (4 / 3)
qqplot_pvals(pvals,
  controls = 500, cases = 499,
  aes_opts = list(null_alpha = 0.5, fit_alpha = 0.5)
)
qqplot_pvals(pvals,
  labels = c("a", "b", "c"),
  controls = 500, cases = 499,
  lambda_correct = TRUE
)
}
