% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{prep}
\alias{bake}
\alias{tidy}
\alias{tunable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tidy}}, \code{\link[generics]{tunable}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{recipes}{\code{\link[recipes]{bake}}, \code{\link[recipes]{prep}}}
}}

