% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-select-genes.R
\name{step_select_genes}
\alias{step_select_genes}
\title{Gene selection by differential expression analysis.}
\usage{
step_select_genes(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  condition = NULL,
  genes_pass = NULL,
  padj_cutoff = 0.05,
  max_n_genes = NULL,
  min_n_genes = NULL,
  options = list(method = "spearman", padj_method = stats::p.adjust.methods[1]),
  skip = FALSE,
  id = recipes::rand_id("select_genes")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which variables will be
used to compute the components. See \code{\link[recipes:selections]{selections()}} for more details. For the
\code{tidy} method, these are not currently used.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned?. By default, the function assumes that the new principal
component columns created by the original variables will be used as
predictors in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{condition}{The condition for the differential expression. See
\code{\link[=cor_de]{cor_de()}}.}

\item{genes_pass}{This should not be specified by the end user. Information
about which genes do and don't pass the p-value threshold in the
differential expression analysis is stored here.}

\item{padj_cutoff}{Genes with an adjusted p-value less than or equal to
\code{padj_cutoff} in the differential expression analysis are kept. The rest
are discarded.}

\item{max_n_genes}{A positive integer. The maximum number of genes selected
by this step.}

\item{min_n_genes}{A positive integer. The minimum number of genes selected.
This guarantees that even if no genes pass \code{padj_cutoff}, there will be
this many returned.}

\item{options}{A list with two elements named \code{method} and \code{padj_method}.
Both are passed to \code{\link[=cor_de]{cor_de()}}. \code{padj_cutoff} is the threshold for keeping
genes.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake.recipe()}}? While all operations are baked
when \code{\link[recipes:prep]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of recipe with the new step added to the sequence
of existing steps (if any).
}
\description{
\code{step_select_genes()} creates a \emph{specification} of a recipe step that will
select genes by differential expression analysis, discarding those that don't
pass a certain p-value threshold. Currently, \code{\link[=cor_de]{cor_de()}} is used.
}
