% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-lm.R
\name{summary.mirmodels_multi_aov_df}
\alias{summary.mirmodels_multi_aov_df}
\title{Summarize the output of \code{\link[=multi_aov]{multi_aov()}}.}
\usage{
\method{summary}{mirmodels_multi_aov_df}(object, ..., na_rm = TRUE)
}
\arguments{
\item{object}{The output of \code{\link[=multi_aov]{multi_aov()}}.}

\item{...}{Optional. Positive integers. Quantiles of variance explained.
See example in \code{\link[=multi_aov]{multi_aov()}}.}

\item{na_rm}{A flag. Remove \code{NA}s in the summary calculations?}
}
\value{
A data frame.
}
\description{
Group by \code{x} and get the mean and median percent variance explained for all
of the \code{y}s.
}
