% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_mirvie_gbm}
\alias{use_mirvie_gbm}
\title{Set up a GBM.}
\usage{
use_mirvie_gbm(mode, outcome)
}
\arguments{
\item{mode}{A string. \code{"classification"} or \code{"regression"}.}

\item{outcome}{A string. The name of the outcome variable in your data.}
}
\value{
The character vector copied to clipboard, invisibly.
}
\description{
Copy boilerplate for fitting a GBM to the clipboard.
}
