% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_mirvie_glm}
\alias{use_mirvie_glm}
\title{Set up a GLM.}
\usage{
use_mirvie_glm(outcome)
}
\arguments{
\item{outcome}{A string. The name of the outcome variable in your data.}
}
\value{
The character vector copied to clipboard, invisibly.
}
\description{
Copy boilerplate for fitting a GLM to the clipboard.
}
