% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-importance.R
\name{vimp}
\alias{vimp}
\title{Get variable importances from a model or list of models.}
\usage{
vimp(mod)
}
\arguments{
\item{mod}{A \code{workflow} or \code{model_fit} object, or a list thereof.}
}
\value{
A tibble.
}
\description{
This is a wrapper around \code{\link[vip:vi]{vip::vi()}}. It returns variables in descending
order of importance. For a list of models, it assumes that the models have
all modelled the same problem with the same variables and the variables are
summarized by median importance.
}
\examples{
iris_data <- janitor::clean_names(datasets::iris)
iris_data_split <- rsample::initial_split(iris_data, strata = species)
mod <- train_glm(
  training_data = rsample::training(iris_data_split),
  outcome = "species",
  metric = "mn_log_loss",
  cv_nfolds = 5,
  n_cores = 5
)
vimp(mod)
vimp(list(mod, mod))
}
